moved and updated menu bar to match latest figma
