+++ src/components/Button.tsx
@@ -160,7 +160,6 @@ const styles = StyleSheet.create({
   },
   content: {
     textAlign: 'left',
-    marginRight: 20,
   },
   chevronOffset: {
     position: 'absolute',
